feat(api): public REST endpoint for action schemas
REST clients had no way to introspect plugin action schemas without
going through MCP. The MCP list_action_schemas tool exists, but a
parallel REST entry point did not — every REST integrator was forced
into MCP for schema discovery, and that gap is the upstream cause of a
class of silent-config bugs (workflow PATCHes accepted with unknown
actionTypes or invalid configs because the caller could not validate
locally).

Adds:

- GET /api/action-schemas — bearer kh_ auth, returns the same payload
  shape as list_action_schemas. Supports ?category=, ?includeChains=,
  and ?type=<actionType> to narrow the actions map to a single entry.
- GET /api/action-schemas/{actionType} — bearer kh_ auth, returns just
  the matching action schema; 404 on unknown actionType.

Extracts the response builder out of /api/mcp/schemas/route.ts into
lib/action-schemas/builder.ts so the MCP-internal route and the new
public REST route share a single source of truth and cannot drift.
/api/mcp/schemas behaviour, response keys, and cache headers are
unchanged.
